feat: check_valid_permission in role_sio & test_role, change setup.cfg & fixed little bugs

+++ xieffect/communities/base/role_sio.py
@@ -6,7 +6,13 @@ from pydantic import BaseModel, Field
 
 from common import EventController, db
 from .meta_db import Community, ParticipantRole
-from .role_db import Role, RolePermission, PermissionTypes, LIMITING_QUANTITY_ROLES
+from .role_db import (
+    Role,
+    RolePermission,
+    PermissionTypes,
+    LIMITING_QUANTITY_ROLES,
+    validation_permissions,
+)
 from ..utils import check_participant
 
 controller = EventController()
@@ -54,11 +60,14 @@ class RolesEventSpace(EventSpace):
         role = Role.create(name=name, color=color, community_id=community.id)
 
         if permissions is not None:
-            for permission in permissions:
-                RolePermission.create(
-                    role_id=role.id,
-                    permission_type=PermissionTypes.from_string(permission),
-                )
+            if validation_permissions(permissions) is not None:
+                for permission in permissions:
+                    RolePermission.create(
+                        role_id=role.id,
+                        permission_type=PermissionTypes.from_string(permission),
+                    )
+            else:
+                controller.abort(400, "Permissions aren't correct")
 
         db.session.commit()
         event.emit_convert(role, self.room_name(community.id))
@@ -86,12 +95,15 @@ class RolesEventSpace(EventSpace):
         if color is not None:
             role.color = color
         if permissions is not None:
-            RolePermission.delete_by_role(role_id=role.id)
-            for permission in permissions:
-                RolePermission.create(
-                    role_id=role.id,
-                    permission_type=PermissionTypes.from_string(permission),
-                )
+            if validation_permissions(permissions) is not None:
+                RolePermission.delete_by_role(role_id=role.id)
+                for permission in permissions:
+                    RolePermission.create(
+                        role_id=role.id,
+                        permission_type=PermissionTypes.from_string(permission),
+                    )
+        else:
+            controller.abort(400, "Permissions aren't correct")
         db.session.commit()
         event.emit_convert(role, self.room_name(community.id))
         return role
